Make plugin compatible with other plugin managers
I'd like to add this to [awesome-zsh-plugins](https://github.com/unixorn/awesome-zsh-plugins), but as written, it will only work with oh-my-zsh because this line in the plugin file assumes the plugin is in the stock oh-my-zsh plugin location. 

`source ~/.oh-my-zsh/plugins/zsh-toolbox/config.zsh`

It also won't work if someone moves their oh-my-zsh directory.